feat(frontend/git): add resizable commit history panel
- Add resizable history panel with drag handle
- Extract time formatting logic to separate utility module
- Update commit list layout and styling
- Improve scroll behavior for commit history
